# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Yral AI Chat API
  description: AI Chat API for Yral with multimodal support
  version: 1.0.0
paths:
...
  /status:
    get:
      tags:
      - Health
      summary: System status
      description: Get detailed system statistics including database info, uptime,
        and system health
      operationId: systemStatus
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '500':
          description: Internal server error
...
components:
  schemas:
...
    DatabaseStats:
      properties:
        connected:
          type: boolean
          title: Connected
        pool_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pool Size
        active_connections:
          anyOf:
          - type: integer
          - type: 'null'
          title: Active Connections
      type: object
      required:
      - connected
      title: DatabaseStats
      description: Database statistics
...
    StatusResponse:
      properties:
        service:
          type: string
          title: Service
        version:
          type: string
          title: Version
        environment:
          type: string
          title: Environment
        uptime_seconds:
          type: integer
          title: Uptime Seconds
        database:
          $ref: '#/components/schemas/DatabaseStats'
        statistics:
          $ref: '#/components/schemas/SystemStatistics'
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - service
      - version
      - environment
      - uptime_seconds
      - database
      - statistics
      - timestamp
      title: StatusResponse
      description: System status response
    SystemStatistics:
      properties:
        total_conversations:
          type: integer
          title: Total Conversations
          default: 0
        total_messages:
          type: integer
          title: Total Messages
          default: 0
        active_influencers:
          type: integer
          title: Active Influencers
          default: 0
      type: object
      title: SystemStatistics
      description: System-wide statistics
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token (without 'Bearer' prefix)